requirements.txt
openai>=1.10.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.0.0
